refactor: modularize kanban board
